+++ experiments/common.py
@@ -1,3 +1,5 @@
+import logging
+import torch
 import math
 import numpy
 from time import sleep
@@ -6,6 +8,7 @@ from time import sleep
 passableBlocks = ['air', 'water', 'lava', 'double_plant', 'tallgrass', 
                   'reeds', 'red_flower', 'yellow_flower', 'flowing_lava',
                   'cobblestone', 'stone', 'sandstone', 'lapis_block']
+
 
 block_id_cliff_walking = {'air': 0,
             'water': 1,
@@ -112,3 +115,28 @@ def direction_to_target(mc, pos):
     dist = math.sqrt((aPos[0] - pos[0]) * (aPos[0] - pos[0]) + (aPos[2] - pos[2]) * (aPos[2] - pos[2]))
     return pitch, yaw, dist
 
+
+def stop_motion(mc):
+    mc.sendCommand('move 0')
+    mc.sendCommand('strafe 0')
+    mc.sendCommand('pitch 0')
+    mc.sendCommand('turn 0')
+    mc.sendCommand('jump 0')
+
+
+def learn(agent, optimizer):
+    losses = []
+    for i in range(40):
+        optimizer.zero_grad()
+        loss = agent.compute_loss()
+        if loss is not None:
+            # Optimize the model
+            loss.backward()
+            torch.nn.utils.clip_grad_norm_(agent.parameters(), 2)
+            optimizer.step()
+            losses.append(loss.cpu().detach())
+    if losses:
+        logging.debug('optimizing')
+        logging.debug('loss %f', numpy.mean(losses))
+
+    return numpy.mean(losses)
